Allow field value unset (set to null by Model::setNull())
